retarget: fix main redirection and sdk/mbed init
As mbed 2 lib works without rtos, we have to have a baremetal boot as it used to
be in mbed_retarget. We create new mbed_main.c that contains main wrapping and
calls prior main (mbed sdk init and mbed main).

This should work without RTOS and with RTOS1/RTOS2.

Rework mbed main and RTX_CM_Lib (these changees used to be in retarget file).
mbed_main is used for mbed 2. It is used when there's no RTOS config present,
in our case it is for old RTX4. This should bring backward compatibility (no breaking change for the mbed 2 users).

Note: The mbed OS has own boot unified process