Add codeowners file
Defining code owners the code within mbed-os repository. As code-owners need write
access to the repository, this makes it very limited in our case. We should use bot
instead to read this file, and either @mention the owners in the PR or if possible
assign them as reviewers.

For more info https://help.github.com/articles/about-codeowners/