CMake: Have only one CMake target for each component
Remove component targets with board, manufacturer suffixes
to simply the user experience. The board specific files
are automatically added to the build based on the labels
of the mbed targets without using the custom utility function.
The utility function was not used as it relies on the
previous tools naming convention for including directories.

Includes: nfc, 802.15.4_RF, cellular