CMake ble: add a comment why drivers are INTERFACE lib
If it is an OBJECT it would need mbed-ble and mbed-core. As we want to link these drivers to
mbed-ble to be available for an app (the only required step for an app is to link to mbed-ble and
we figure out internally what is required for a specific board - using config library add/remove).

This can be refactored once we get CMake config story to replace our config (or at least targets.json information like add/remove libraries).